Execute examples from Rd files as testthat tests

Reads examples from Rd files and constructs testthat-style tests. testthat expectations are built such that

test_examples_as_testthat(
  package,
  path,
  ...,
  test_dir = tempfile("testex"),
  clean = TRUE,
  overwrite = TRUE,
  reporter = testthat::get_reporter()
)

Arguments

package

A package name whose examples should be tested

path

Optionally, a path to a source code directory to use. Will only have an effect if parameter package is missing.

...

Additional argument unused

test_dir

An option directory where test files should be written. Defaults to a temporary directory.

clean

Whether the test_dir should be removed upon completion of test execution. Defaults to TRUE.

overwrite

Whether files should be overwritten if test_dir already exists. Defaults to TRUE.

reporter

A testthat reporter to use. Defaults to the active reporter in the testthat environment or default reporter.

Details

Each example expression is expected to run without error

Any testex expectations are expected to pass