Support for testthat example expectations

Various functions that are used to produce a more native testthat experience, automatically converting testex tests into testthat code and executing tests such that they produce informative messages on failure.

testthat_block(
  ...,
  value = get_example_value(),
  obj = NULL,
  example = NULL,
  tests = NULL,
  envir = parent.frame()
)

with_srcref(src, expr, envir = parent.frame())

expect_no_error(object, ...)

Arguments

...

Additional arguments unused

value

A symbol or quote to use to refer to the subject of testthat tests.

obj

An optional object name used to construct a more helpful error message testthat failure message.

example

An option srcref_key string used to indicate where the relevant example code originated from.

tests

An option srcref_key string used to indicate where the relevant test code originated from.

envir

An environment in which the expectations should be evaluated

src

A srcref_key which is parsed to produce an artificial srcref for the expectation signaled messages.

expr

An expression to be evaluated. If an expectation condition is raised during its evaluation, its srcref is converted to src.

object

An expression to evaluate

Details

testex operates on the previous value produced in example code. This is unlike testthat expectations, which expect a value to be provided as a first argument.

To accommodate a more native testthat interface, testex provides a few convenience functions to make testex expectations run more natively within the style of testthat.

Functions

testthat_block(): A flavor of testex that will inject .Last.value into the first argument of each expression - suitable for using the expect_* family of functions from testthat. Also handles temporarily attaching the testthat package.

with_srcref(): Retroactively assigns a source file and location to a expectation. This allows testthat to report an origin for any code that raised an example test failure from the source roxygen code, even though the test code is reconstructed from package documentation files.

expect_no_error(): A testthat expectation that the provided code can be evaluated without producing an error. This is the most basic expectation one should expect of any example code. Further expectations are provided in subsequent testthat code.

Examples