Add testex tags and configure package to fully use testex features

Add testex tags and configure package to fully use testex features

Usage

use_testex(path = getwd(), check = TRUE, quiet = FALSE)

Arguments

path

A package source code working directory

check

A logical value indicating whether tests should be executing during R CMD check.

quiet

Whether output should be suppressed

Value

The result of write.dcf() upon modifying the package DESCRIPTION file.

Note

The testex roxygen2 tags behave similarly to roxygen2 @examples tags, with the minor addition of some wrapping code to manage the tests. This means that they will be integrated into your @examples and can be intermixed between @examples tags